Salim

+++ python_async_function/4-tasks.py
@@ -7,7 +7,7 @@ from typing import List
 task_wait_random = __import__('3-tasks').task_wait_random
 
 
-async def task_wait_n(n: int, max_delay: int) -> List[int]:
+async def task_wait_n(n: int, max_delay: int) -> List[float]:
     """Spawns task_wait_random n times and returns delays in ascending order"""
 
     tasks = [task_wait_random(max_delay) for i in range(n)]
